requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
firecrawl-py==1.2.4
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.5.0
loguru==0.7.2
multidict==6.1.0
nest-asyncio==1.6.0
openai==1.51.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
setuptools==75.1.0
sniffio==1.3.1
tqdm==4.66.5
twilio==9.3.3
typing_extensions==4.12.2
urllib3==2.2.3
websockets==13.1
wheel==0.44.0
yarl==1.13.1
